requirements.txt
aiogram>=3.0.0
python-dotenv
aiosqlite
aiohttp
requests
